eu_open_data_api/requirements.txt
fastapi
uvicorn
httpx  # Para hacer peticiones HTTP a la API externa
pymongo  # Para interactuar con MongoDB
python-dotenv  # Para cargar variables de entorno desde .env
